LayoutTests/inspector-protocol/input/dispatchMouseEvent-expected.txt

-----Event-----
type: mousedown
button: 0
x: 100
y: 200
-----Event-----
type: mouseup
button: 0
x: 100
y: 200
-----Event-----
type: mousemove
button: 0
shiftKey
x: 50
y: 150
